updated backend directory structure for more modularity
