+++ backend/server.js
@@ -1,67 +1,12 @@
-// server.js
-const dotenv = require('dotenv');
-const express = require('express');
-const cors = require('cors');
-const { v4: uuidv4 } = require('uuid');
+const app = require('./src/app');  // Importing app with all middleware
 
-dotenv.config();
+// Uncomment if we need custom env
+// const dotenv = require('dotenv');
 
-// Initialize express
-const app = express();
-const PORT = process.env.PORT || 5000; // Use an environment variable or default to 5000
+// dotenv.config();
 
-// Middleware
-app.use(cors());
-app.use(express.json()); // Parse JSON bodies (for POST/PUT requests)
+const PORT = process.env.PORT || 5000;
 
-// In-memory storage for sessions (for now, but consider using a database later)
-const sessions = {};
-
-// Start new game session
-app.post('/reaction/start', (req, res) => {
-  console.log("\n\nReaction Game Start")
-  const sessionId = uuidv4();
-
-  sessions[sessionId] = {
-    bestReactionTime: null
-  };
-
-  return res.json({ sessionId });
-});
-
-// End game session. Frontend calculates the reaction time -- lower latency
-app.post('/reaction/end', (req, res) => {
-  console.log("\n\nReaction Game End")
-
-  const { sessionId, reactionTime } = req.body;
-
-  if (!sessionId || !reactionTime) {
-    return res.status(400).json({ message: 'Session ID and reaction time are required.' });
-  }
-
-  const session = sessions[sessionId];
-
-  if (!session) {
-    return res.status(404).json({ message: 'Session not found. Start game first.' });
-  }
-
-  if (session.bestReactionTime === null || reactionTime < session.bestReactionTime) {
-    session.bestReactionTime = reactionTime;
-  }
-
-  return res.json({ 
-    message: 'Reaction time recorded.', 
-    bestReactionTime: session.bestReactionTime 
-  });
-});
-
-
-// Hello world endpoint
-app.get('/', (req, res) => {
-  res.send('Hello from the backend!');
-});
-
-// Start the server
 app.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}`);
+  console.log(`Server running on port ${PORT}`);
 });
