+++ backend/src/config/db.js
@@ -1,0 +1,1 @@
+// TODO, connect to a MySQL DB
